% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habitat.R
\name{add_habitat}
\alias{add_habitat}
\title{Add habitat zones to a Habitat object}
\usage{
add_habitat(hab, sf_data, weight = 1, ...)
}
\arguments{
\item{hab}{A Habitat object}

\item{sf_data}{An sf object containing geometries to add}

\item{weight}{A weight to apply on the habitat feature. Can then be used as
resistance for dispersal functions for instance.}

\item{...}{Additional arguments}
}
\value{
A Habitat object with new geometries added as habitat
}
\description{
Add habitat zones to a Habitat object
}
