% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habitat.R
\name{add_nohabitat}
\alias{add_nohabitat}
\title{Add non-habitat zones to a Habitat object}
\usage{
add_nohabitat(hab, sf_data, ...)
}
\arguments{
\item{hab}{A Habitat object}

\item{sf_data}{An sf object containing geometries to add}

\item{...}{Additional arguments}
}
\value{
A Habitat object with new geometries added as non-habitat
}
\description{
Add non-habitat zones to a Habitat object
}
