% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bappet_cd}
\alias{bappet_cd}
\title{Data concentration Cd Soil - Veetation extracted from BAPPET}
\format{
An object of class \code{data.frame} with 3040 rows and 80 columns.
}
\usage{
data(bappet_cd)
}
\description{
Data concentration Cd Soil - Veetation extracted from BAPPET
}
\examples{
data(bappet_cd)
}
\keyword{datasets}
