% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel.R
\name{compute_kernel}
\alias{compute_kernel}
\title{Create a 2D Gaussian motion kernel as a SpatRaster}
\usage{
compute_kernel(radius, GSD, size_std = 1.5)
}
\arguments{
\item{radius}{Numeric, std of the distribution in meters}

\item{GSD}{Numeric, ground sampling distance in meters per pixel}

\item{size_std}{Numeric, how many std to extend kernel on each side}
}
\value{
SpatRaster of the kernel
}
\description{
Create a 2D Gaussian motion kernel as a SpatRaster
}
