% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DBFunc_BirdFuncDat}
\alias{DBFunc_BirdFuncDat}
\title{Bird Functional Traits Dataset (EltonTraits 1.0)}
\format{
A data frame containing species dietary breakdowns (summing to 100%) and trait data:
\describe{
  \item{SpecID}{Unique species identifier.}
  \item{PassNonPass}{Indicates whether the bird is a Passerine or Non-Passerine.}
  \item{IOCOrder}{IOC taxonomic order.}
  \item{BLFamilyLatin}{BirdLife International family Latin name.}
  \item{BLFamilyEnglish}{BirdLife International family English name.}
  \item{BLFamSequID}{BirdLife International family sequence ID.}
  \item{Taxo}{Taxonomic grouping code.}
  \item{Scientific}{Scientific name of the bird species.}
  \item{English}{Common English name of the bird species.}
  \item{Diet.Inv}{Percentage of invertebrates in the diet.}
  \item{Diet.Vend}{Percentage of vertebrate endotherms consumed.}
  \item{Diet.Vect}{Percentage of vertebrate ectotherms consumed.}
  \item{Diet.Vfish}{Percentage of fish consumed.}
  \item{Diet.Vunk}{Percentage of unknown/unclassified vertebrates in the diet.}
  \item{Diet.Scav}{Percentage of scavenging activity.}
  \item{Diet.Fruit}{Percentage of fruits consumed.}
  \item{Diet.Nect}{Percentage of nectar/pollen consumed.}
  \item{Diet.Seed}{Percentage of seeds/nuts consumed.}
  \item{Diet.PlantO}{Percentage of other plant tissues consumed.}
  \item{Diet.5Cat}{Simplified 5-category diet classification.}
  \item{Diet.Source}{Source of the dietary information.}
  \item{Diet.Certainty}{Certainty score for the diet data.}
  \item{Diet.EnteredBy}{Identifier of the person who entered the diet data.}
  \item{ForStrat.watbelowsurf}{Percentage of foraging time spent underwater.}
  \item{ForStrat.wataroundsurf}{Percentage of foraging time spent at the water surface.}
  \item{ForStrat.ground}{Percentage of foraging time spent on the ground.}
  \item{ForStrat.understory}{Percentage of foraging time spent in the understory.}
  \item{ForStrat.midhigh}{Percentage of foraging time spent in the mid-high canopy.}
  \item{ForStrat.canopy}{Percentage of foraging time spent in the upper canopy.}
  \item{ForStrat.aerial}{Percentage of foraging time spent in aerial foraging.}
  \item{PelagicSpecialist}{Indicator of pelagic specialization.}
  \item{ForStrat.Source}{Source of the foraging strategy data.}
  \item{ForStrat.SpecLevel}{Specificity level of the foraging strategy record.}
  \item{ForStrat.EnteredBy}{Identifier of the person who entered the foraging data.}
  \item{Nocturnal}{Indicator of nocturnal activity.}
  \item{BodyMass.Value}{Body mass value in grams.}
  \item{BodyMass.Source}{Source of the body mass information.}
  \item{BodyMass.SpecLevel}{Specificity level of the body mass record.}
  \item{BodyMass.Comment}{Notes on body mass data.}
  \item{Record.Comment}{General comments about the record.}
}
}
\source{
EltonTraits 1.0 database.
}
\usage{
DBFunc_BirdFuncDat
}
\description{
Dietary category percentages, foraging strategies, and functional traits for birds.
}
\keyword{datasets}
