% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DBFunc_MamFuncDat}
\alias{DBFunc_MamFuncDat}
\title{Mammal Functional Traits Dataset (EltonTraits 1.0)}
\format{
A data frame containing species dietary breakdowns (summing to 100%) and trait data:
\describe{
  \item{MSW3_ID}{Mammal Species of the World (3rd ed.) identifier.}
  \item{Scientific}{Scientific name of the mammal species.}
  \item{MSWFamilyLatin}{Mammal Species of the World (3rd ed.) family name.}
  \item{Diet.Inv}{Percentage of invertebrates in the diet.}
  \item{Diet.Vend}{Percentage of vertebrate endotherms consumed (birds and mammals).}
  \item{Diet.Vect}{Percentage of vertebrate ectotherms consumed (reptiles and amphibians).}
  \item{Diet.Vfish}{Percentage of fish consumed.}
  \item{Diet.Vunk}{Percentage of unknown/unclassified vertebrates in the diet.}
  \item{Diet.Scav}{Percentage of scavenging activity.}
  \item{Diet.Fruit}{Percentage of fruits consumed.}
  \item{Diet.Nect}{Percentage of nectar/pollen consumed.}
  \item{Diet.Seed}{Percentage of seeds/nuts consumed.}
  \item{Diet.PlantO}{Percentage of other plant tissues consumed (leaves, stems, roots).}
  \item{Diet.Source}{Source of the dietary information.}
  \item{Diet.Certainty}{Certainty score for the diet data.}
  \item{ForStrat.Value}{Primary foraging stratum/habitat.}
  \item{ForStrat.Certainty}{Certainty score for the foraging stratum.}
  \item{ForStrat.Comment}{Notes on foraging strategy.}
  \item{Activity.Nocturnal}{Binary flag for nocturnal activity.}
  \item{Activity.Crepuscular}{Binary flag for crepuscular activity.}
  \item{Activity.Diurnal}{Binary flag for diurnal activity.}
  \item{Activity.Source}{Source of the activity information.}
  \item{Activity.Certainty}{Certainty score for the activity pattern.}
  \item{BodyMass.Value}{Body mass value in grams.}
  \item{BodyMass.Source}{Source of the body mass information.}
  \item{BodyMass.SpecLevel}{Specificity level of the body mass record.}
}
}
\source{
EltonTraits 1.0 database.
}
\usage{
DBFunc_MamFuncDat
}
\description{
Dietary category percentages, foraging strategies, and functional traits for mammals.
}
\keyword{datasets}
