% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispersal.R
\name{dispersal}
\alias{dispersal}
\title{Create a raster mask from a habitat object}
\usage{
dispersal(
  spacemodel,
  layer = 1,
  method = "convolution",
  method_option = list()
)

dispersal(
  spacemodel,
  layer = 1,
  method = "convolution",
  method_option = list()
)
}
\arguments{
\item{spacemodel}{spacemodel object.}

\item{layer}{character or integer. Name or index of the layer to disperse.}

\item{method}{method use for dispersal ("convolution", etc.).}

\item{method_option}{parameters list (must contain `kernel` for convolution).}
}
\value{
spacemod with modification of the layer

spacemod with modification of the layer
}
\description{
Applies dispersal logic to a specific layer of the spacemodel.
}
