% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{earthworm_cd}
\alias{earthworm_cd}
\title{Data concentration Cd Soil - Earthworm}
\format{
An object of class \code{data.frame} with 91 rows and 4 columns.
}
\usage{
data(earthworm_cd)
}
\description{
Data concentration Cd Soil - Earthworm
}
\examples{
data(earthworm_cd)
}
\keyword{datasets}
