% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flux.R
\name{flux}
\alias{flux}
\title{Constructor for Flux Parameters}
\usage{
flux(x, default = 1, normalize = TRUE)
}
\arguments{
\item{x}{A `spacemodel` object or a `trophic_tbl`.}

\item{default}{The default function or coefficient for unspecified links (default is 1, meaning full transfer: f(x) = x).}

\item{normalize}{Logical. Whether to normalize diet weights (default TRUE).}
}
\value{
A `trophic_tbl` with an initialized `flux` column.
}
\description{
Initializes the trophic flux table for a spacemodel.
}
