% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_extdata.R
\name{load_raster_extdata}
\alias{load_raster_extdata}
\title{load_raster_extdata}
\format{
An object `SpatRaster` (package terra).
}
\source{
Internal data package.
}
\usage{
load_raster_extdata(path_name)
}
\arguments{
\item{path_name}{Path of the raster (.tiff) file to download.}
}
\description{
A raster of the (example).
}
\keyword{datasets}
