% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trophic.R
\name{lower_neighbors}
\alias{lower_neighbors}
\title{Get resource layers for a given trophic layer}
\usage{
lower_neighbors(trophic_tbl, layer)
}
\arguments{
\item{trophic_tbl}{A \code{trophic_tbl} object}

\item{layer}{Name of the layer (string) in the spacemodel.}
}
\value{
A character vector of names of layers that are resources for \code{layer}.
}
\description{
Returns the upstream neighbors (prey/resources) of a given layer in a trophic graph.
}
