% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ocsge_metaleurop}
\alias{ocsge_metaleurop}
\title{SF object defining very simplified OCS-GE soil cover metaleurop}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 9 rows and 12 columns.
}
\usage{
data(ocsge_metaleurop)
}
\description{
Simple feature collection with 9 features and 11 fields.
Projected CRS: RGF93 v1 / Lambert-93.
}
\examples{
data(ocsge_metaleurop)
}
\keyword{datasets}
