% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kriging.R
\name{plot.soil_vgm}
\alias{plot.soil_vgm}
\title{Plot an empirical variogram and its fitted model(s)}
\usage{
\method{plot}{soil_vgm}(x, all_models = TRUE, ...)
}
\arguments{
\item{x}{A `soil_vgm` object (output of [fit_soil_variogram()]).}

\item{all_models}{Logical. Draw every model of `x$comparison`? (default `TRUE`)}

\item{...}{Not used.}
}
\value{
A `ggplot` object.
}
\description{
Plot an empirical variogram and its fitted model(s)
}
