% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ocsge.R
\name{plot_species_habitat}
\alias{plot_species_habitat}
\title{Plot Habitat Suitability Maps from OCSGE Data}
\usage{
plot_species_habitat(sf_merged, title_species = NULL)
}
\arguments{
\item{sf_merged}{An `sf` object enriched with species traits (e.g., output of \code{\link{join_ocsge_species}}).}

\item{title_species}{Character (optional). The name of the species to display in the title.
If NULL, it will try to retrieve the name from the object attributes.}
}
\value{
A \code{patchwork} object containing the combined \code{ggplot2} maps.
}
\description{
Generates a composite plot of three spatial maps (Global Weight, Foraging Weight,
and Resistance) from an `sf` object previously enriched with species traits.
}
