% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ref_ocsge}
\alias{ref_ocsge}
\title{Nomenclature of OCS-GE soil cover}
\format{
An object of class \code{data.frame} with 14 rows and 4 columns.
}
\usage{
data(ref_ocsge)
}
\description{
Nomenclature of OCS-GE soil cover
}
\examples{
data(ref_ocsge)
}
\keyword{datasets}
