% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{roi_metaleurop}
\alias{roi_metaleurop}
\title{SF object defining ROI metaleurop}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 1 rows and 2 columns.
}
\usage{
data(roi_metaleurop)
}
\description{
Simple feature collection with 1 feature and 1 field.
Geodetic CRS:  WGS 84.
}
\examples{
data(roi_metaleurop)
}
\keyword{datasets}
