% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sf_micromammals}
\alias{sf_micromammals}
\title{DataBase of collected MicroMammals species}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1426 rows and 27 columns.
}
\usage{
data(sf_micromammals)
}
\description{
DataBase of collected MicroMammals species
}
\examples{
data(sf_micromammals)
}
\keyword{datasets}
