% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_soil.R
\docType{data}
\name{soil_metaleurop}
\alias{soil_metaleurop}
\title{Soil metal concentrations sampled around the former Metaleurop smelter}
\format{
An `sf` object with 595 rows and 9 columns:
\describe{
  \item{id}{Sample identifier.}
  \item{usage}{Land-use class of the sampled plot, as coded in the original
    survey (`"A"`, `"L"`, `"U"`).}
  \item{dist}{Distance (m) to the smelter chimney.}
  \item{angle}{Direction of the sample seen from the smelter (degrees
    clockwise from the North).}
  \item{wind}{Wind-rose frequency (in \%) of the 20-degree sector containing
    `angle` (18 sectors centred on 0, 20, ..., 340 degrees, summing to 100).
    The highest values lie north-east of the smelter, i.e. downwind of the
    dominant south-westerly winds.}
  \item{cd, pb, zn}{Total concentrations in mg/kg of dry soil.}
  \item{geometry}{Sampling location.}
}
}
\usage{
data(soil_metaleurop)
}
\description{
Topsoil concentrations of cadmium (Cd), lead (Pb) and zinc (Zn) measured at
595 sampling points around the former Metaleurop Nord lead and zinc smelter
(Noyelles-Godault, northern France). This is the data set used to build the
soil rasters shipped in `inst/extdata` (`ground_concentration_*.tif`), see
the kriging article.

Simple feature collection (POINT), projected CRS: RGF93 v1 / Lambert-93
(EPSG:2154).
}
\examples{
data(soil_metaleurop)
summary(soil_metaleurop[, c("cd", "pb", "zn")])
}
\keyword{datasets}
