% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacemodel.R
\name{spacemodel}
\alias{spacemodel}
\title{Create a spacemodel object}
\usage{
spacemodel(raster_stack, trophic_tbl)
}
\arguments{
\item{raster_stack}{SpatRaster (multi-layer stack)}

\item{trophic_tbl}{Object of class \code{trophic_tbl}}
}
\description{
Create a spacemodel object
}
