% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacemodR-internal.R
\name{spacemodR-package}
\alias{spacemodR-package}
\alias{spacemodR}
\title{The 'spacemodR' package.}
\description{
Workflow for Environmental Risk Assessment: Habitat,
Food Web, Dispersal, Exposure and Risk. A set of function for
}
\references{
Stan Development Team (2020). RStan: the R interface to Stan. R package version 2.21.2. https://mc-stan.org
}
