% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trophic.R
\name{[.trophic_tbl}
\alias{[.trophic_tbl}
\title{Subset method for trophic_tbl}
\usage{
\method{[}{trophic_tbl}(x, ...)
}
\arguments{
\item{x}{A trophic_tbl object.}

\item{...}{Additional arguments passed to the base method.}
}
\value{
A validated trophic_tbl object.
}
\description{
Ensures that any subsetting or modification preserves the validity
of the trophic table.
}
