% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trophic.R
\name{validate_trophic_tbl}
\alias{validate_trophic_tbl}
\title{Validate a trophic table}
\usage{
validate_trophic_tbl(df)
}
\arguments{
\item{df}{A data.frame intended to be a trophic_tbl object.}
}
\value{
A validated \code{trophic_tbl} object.
}
\description{
Internal function used to ensure that a trophic_tbl object respects
all structural constraints.
}
\details{
The function checks that:
\itemize{
  \item All links are unique
  \item No self-loops are present
  \item The graph is acyclic
}
It also computes and updates the trophic level attribute.
}
\keyword{internal}
