% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ocsge.R
\name{get_ocsge_data_fgb}
\alias{get_ocsge_data_fgb}
\title{Collect data from OCS GE repository using FlatGeobuf on S3}
\usage{
get_ocsge_data_fgb(roi, fgb_url)
}
\arguments{
\item{roi}{An object `sf` defiing Region Of Interest.}

\item{fgb_url}{Public URL or a .fgb file.}
}
\value{
An object `sf` cropped to the bounding bbox form ROI.
}
\description{
Collect data from OCS GE repository using FlatGeobuf on S3
}
