% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habitat.R
\name{habitat_raster}
\alias{habitat_raster}
\title{Create a raster from a habitat object}
\usage{
habitat_raster(ground_raster, habitat)
}
\arguments{
\item{ground_raster}{SpatRaster (reference grid).}

\item{habitat}{habitat object inherited from \code{sf} data.frame class.}
}
\value{
SpatRaster (weight and NA in nogo)
}
\description{
Create a raster from a habitat object
}
