% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ocsge_species_dict}
\alias{ocsge_species_dict}
\title{Valued weight between OCSGE layer and species}
\format{
An object of class \code{data.frame} with 13090 rows and 6 columns.
}
\usage{
data(ocsge_species_dict)
}
\description{
Valued weight between OCSGE layer and species
}
\examples{
data(ocsge_species_dict)
}
\keyword{datasets}
